feat(post): userdetails 인증 기능 추가

+++ src/main/java/com/sparta/newsfeed/service/PostService.java
@@ -4,24 +4,29 @@ import com.sparta.newsfeed.dto.PostRequsetDto;
 import com.sparta.newsfeed.dto.PostResponseDto;
 import com.sparta.newsfeed.entity.Post;
 import com.sparta.newsfeed.repository.PostRepository;
+import com.sparta.newsfeed.user.entity.User;
+import com.sparta.newsfeed.user.security.UserDetailsImpl;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.RejectedExecutionException;
 
 @Service
 public class PostService {
 
     private final PostRepository postRepository;
-    public PostService(PostRepository postRepository) {
+    private final CommentService commentService;
+    public PostService(PostRepository postRepository, CommentService commentService) {
         this.postRepository = postRepository;
+        this.commentService = commentService;
     }
 
     //게시물 생성
-    public PostResponseDto createPost(PostRequsetDto requsetDto) {
-        Post post = new Post(requsetDto);
+    public PostResponseDto createPost(PostRequsetDto requsetDto, UserDetailsImpl userDetails) {
+        Post post = new Post(requsetDto, userDetails);
         postRepository.save(post);
         return new PostResponseDto(post);
     }
@@ -45,16 +50,40 @@ public class PostService {
     }
     //게시물 수정
     @Transactional
-    public PostResponseDto updatePost(Long postid, PostRequsetDto requsetDto) {
+    public PostResponseDto updatePost(Long postid, PostRequsetDto requsetDto, UserDetailsImpl userDetails) {
         Post post = postRepository.findById(postid).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 id 입니다."));
+
+        if(!userDetails.getUsername().equals(post.getUsername())){
+            throw new RejectedExecutionException("게시물의 작성자만 수정이 가능합니다.");
+        }
         post.update(requsetDto);
         return new PostResponseDto(post);
     }
 
     //게시물 삭제
-    public void deletePost(Long postid) {
+    public void deletePost(Long postid, UserDetailsImpl userDetails) {
         Post post = postRepository.findById(postid).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 id 입니다."));
+
+        if(!userDetails.getUsername().equals(post.getUsername())){
+            throw new RejectedExecutionException("게시물의 작성자만 삭제가 가능합니다.");
+        }
+        commentService.deleteCommentByPostId(post.getId());
         postRepository.delete(post);
     }
 
+    //회원별 게시물 조회
+    public List<PostResponseDto> getPostsByUsername(Long userid, UserDetailsImpl userDetails) {
+        String username = null;
+        if (userid.equals(userDetails.getUser().getId())) {
+            username = userDetails.getUser().getUsername();
+        }
+        List<Post> posts = postRepository.findByUsername(username);
+        List<PostResponseDto> responseDtoList = new ArrayList<>();
+
+        for (Post post : posts) {
+            responseDtoList.add(new PostResponseDto(post));
+        }
+
+        return responseDtoList;
+    }
 }
